Moved coercion to own module.
